refactor- duplicate code
