+++ packages/template-retail-react-app/app/components/bonus-product-view-modal/index.jsx
@@ -21,6 +21,7 @@ import {
 } from '@salesforce/retail-react-app/app/components/shared/ui'
 import ProductView from '@salesforce/retail-react-app/app/components/product-view'
 import {useProductViewModal} from '@salesforce/retail-react-app/app/hooks/use-product-view-modal'
+import {useControlledVariations} from '@salesforce/retail-react-app/app/hooks/use-controlled-variations'
 import {useIntl} from 'react-intl'
 import {useShopperBasketsMutationHelper} from '@salesforce/commerce-sdk-react'
 import {useCurrentBasket} from '@salesforce/retail-react-app/app/hooks/use-current-basket'
@@ -60,40 +61,12 @@ const BonusProductViewModal = ({
         }
     }, [product])
 
-    // Controlled variation values state for modal (doesn't use URL params)
-    const [controlledVariationValues, setControlledVariationValues] = React.useState({})
+    // Use custom hook for controlled variation management
+    const {controlledVariationValues, handleVariationChange} = useControlledVariations(safeProduct)
 
     const productViewModalData = useProductViewModal(safeProduct, controlledVariationValues, {
         keepPreviousData: true
     })
-
-    // Auto-select variation attributes with only one value
-    React.useEffect(() => {
-        if (!productViewModalData.product?.variationAttributes) return
-
-        const autoSelections = {}
-        productViewModalData.product.variationAttributes.forEach((attr) => {
-            // Only auto-select if there's exactly one value and it's not already selected
-            if (attr.values?.length === 1 && !controlledVariationValues[attr.id]) {
-                autoSelections[attr.id] = attr.values[0].value
-            }
-        })
-
-        if (Object.keys(autoSelections).length > 0) {
-            setControlledVariationValues((prev) => ({
-                ...prev,
-                ...autoSelections
-            }))
-        }
-    }, [productViewModalData.product?.variationAttributes, controlledVariationValues])
-
-    // Handle variation changes in controlled mode
-    const handleVariationChange = React.useCallback((attributeId, value) => {
-        setControlledVariationValues((prev) => ({
-            ...prev,
-            [attributeId]: value
-        }))
-    }, [])
 
     // Keep a stable reference to the last successfully loaded product
     // This prevents constant re-renders while fetching
